refactor(config): move types and inversify config to config folder for better organization
This commit relocates the `types.ts` and `inversify.config.ts` files to a new `config` folder to improve project structure. Additionally, it updates all import paths to reflect the new locations and introduces a new `@config` path alias in `tsconfig.json`. The changes also include adding a new `fetchAll` method to `ICategoryRepository` and updating the `GetCategoryTreeHandler` to use this method. Swagger and morgan dependencies were added to enhance API documentation and logging capabilities.